process for adding service

+++ gbc_mads_4002_project/FitnessClass.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 class FitnessClass: Service {
-    var duration: Double // in hours
+    var duration: Int // in hours
     var trainerName: String
     
     override func getServiceInfo() -> String {
@@ -18,7 +18,7 @@ class FitnessClass: Service {
     }
     
     //constructor 
-    init (id: Int, name: String, numberOfSessions: Int, price: Int, duration: Double, trainerName: String) {
+    init (id: Int, name: String, numberOfSessions: Int, price: Int, duration: Int, trainerName: String) {
         self.duration = duration
         self.trainerName = trainerName
         super.init(id: id, name: name, numberOfSessions: numberOfSessions, price: price)
