+++ gbc_mads_4002_project/PersonalTraining.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 class PersonalTraining: Service {
-    var sessionTime: Double // in hours
+    var sessionTime: Int // in hours
     
     override func getServiceInfo() -> String {
         return "Personal Training " + super.getServiceInfo() +
@@ -16,7 +16,7 @@ class PersonalTraining: Service {
     }
     
     //constructor 
-    init (id: Int, name: String, numberOfSessions: Int, price: Int, sessionTime: Double) {
+    init (id: Int, name: String, numberOfSessions: Int, price: Int, sessionTime: Int) {
         self.sessionTime = sessionTime
         super.init(id: id, name: name, numberOfSessions: numberOfSessions, price: price)
     }
